requirements.txt
python-telegram-bot
openai
moviepy
elevenlabs
pexels_api
requests
pillow
numpy
pysrt
dropbox
ffmpeg
